+++ GW2EIParser/EIData/Statistics/FinalPlayerSupport.cs
@@ -4,13 +4,13 @@
     public class FinalPlayerSupport
     {
         //public long allHeal;
-        public long Resurrects { get; set; }
+        public int Resurrects { get; set; }
         public double ResurrectTime { get; set; }
-        public long CondiCleanse { get; set; }
+        public int CondiCleanse { get; set; }
         public double CondiCleanseTime { get; set; }
-        public long CondiCleanseSelf { get; set; }
+        public int CondiCleanseSelf { get; set; }
         public double CondiCleanseTimeSelf { get; set; }
-        public long BoonStrips { get; set; }
+        public int BoonStrips { get; set; }
         public double BoonStripsTime { get; set; }
     }
 }
